+++ tools/crypto/sha1.cpp
@@ -32,7 +32,7 @@ void _clp_crypto_sha1(int argno)
 {
     CCC_PROLOG("crypto_sha1",1);
     char *text=_parb(1);
-    int length=_parblen(1);
+    size_t length=_parblen(1);
 
     SHA_CTX *c=new SHA_CTX();
     SHA1_Init(c);
@@ -63,7 +63,7 @@ void _clp_crypto_sha1_update(int argno)
     CCC_PROLOG("crypto_sha1_update",2);
     SHA_CTX *c=(SHA_CTX*)_parp(1);
     char *text=_parb(2);
-    int length=_parblen(2);
+    size_t length=_parblen(2);
     SHA1_Update(c,text,length);
     _ret();
     CCC_EPILOG();
